Place widgets by their edges, not in 576 squares
The board's coordinates were integers because placement was a boolean grid:
`_occupancy` rasterised rows x cols of taken-or-free and everything else walked
it. That is what forced a widget into the nearest of 576 squares rather than
where it was actually put.

Coordinates are now floats on the same 32-by-18 scale, so every call written
before this goes on meaning exactly what it meant, and `placement.py` is
rewritten around rectangles.

The idea that makes it tractable is that you never need every position, only
the ones touching something: a widget that may not overlap always comes to rest
flush against another widget's edge or a wall, so the candidate positions are
just the edges already on the board. `find_slot` walks them in the same reading
order the cell scan used and returns the same slot it did — the sliding
argument in its docstring is why that is not a coincidence. `is_free` drops to
rectangle intersection. `largest_free_rect` stays exact by sweeping the widgets
crossing each band between two horizontal edges.

`ge=1` becomes a real floor rather than zero: MIN_SIZE is a quarter of a cell,
about 15px on the television, below which a widget is not small but invisible
while still refusing to let anything overlap it.

Every measurement a human or a model reads goes through `cells()`, so a board
576 cells across does not start advertising itself as 576.0.

Closes #35.

+++ backend/repositories/board.py
@@ -43,10 +43,10 @@ def get_by_key(key: str) -> ItemRead | None:
 
 def add(
     payload: Payload,
-    x: int,
-    y: int,
-    w: int,
-    h: int,
+    x: float,
+    y: float,
+    w: float,
+    h: float,
     parent_id: str | None,
     pinned: bool,
     key: str | None = None,
